move 'compositesampling' package to 'sampling'
