+++ custom/src/main/java/co/elastic/otel/dynamicconfig/CentralConfig.java
@@ -18,7 +18,7 @@
  */
 package co.elastic.otel.dynamicconfig;
 
-import co.elastic.otel.compositesampling.DynamicCompositeParentBasedTraceIdRatioBasedSampler;
+import co.elastic.otel.sampling.DynamicCompositeParentBasedTraceIdRatioBasedSampler;
 import co.elastic.otel.config.ConfigLoggingAgentListener;
 import co.elastic.otel.dynamicconfig.internal.OpampManager;
 import co.elastic.otel.logging.AgentLog;
