+++ custom/src/main/java/co/elastic/otel/sampling/DynamicCompositeParentBasedTraceIdRatioBasedSampler.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.otel.compositesampling;
+package co.elastic.otel.sampling;
 
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.trace.SpanKind;
